requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy>=2.0.25
alembic==1.12.1
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
Pillow>=11.3.0
bcrypt==4.1.2
python-dateutil==2.8.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-Mail==0.9.1
